requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncspotify==0.12.1
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
frozenlist==1.3.0
idna==3.2
multidict==6.0.2
python-decouple==3.6
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.6
vk-api==11.9.4
yarl==1.7.2
